[#noissue] Fix: Handle when it has a certain base href other than empty string
- Filter the base href on navigation when it's spotted in the delivered url
