+++ web/src/main/webapp/v2/src/app/core/components/configuration-icon/configuration-icon-container.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 
-import { AnalyticsService, TRACKED_EVENT_LIST, DynamicPopupService, StoreHelperService, WindowRefService } from 'app/shared/services';
+import { AnalyticsService, TRACKED_EVENT_LIST, DynamicPopupService } from 'app/shared/services';
 import { ConfigurationPopupContainerComponent } from 'app/core/components/configuration-popup/configuration-popup-container.component';
-import { Actions } from 'app/shared/store';
 
 @Component({
     selector: 'pp-configuration-icon-container',
@@ -14,8 +13,6 @@ export class ConfigurationIconContainerComponent implements OnInit {
     constructor(
         private dynamicPopupService: DynamicPopupService,
         private analyticsService: AnalyticsService,
-        private storeHelperService: StoreHelperService,
-        private windowRefService: WindowRefService,
     ) {}
 
     ngOnInit() {}
@@ -25,12 +22,5 @@ export class ConfigurationIconContainerComponent implements OnInit {
             coord,
             component: ConfigurationPopupContainerComponent
         });
-        this.updateURLPathState();
-    }
-
-    private updateURLPathState(): void {
-        const pathName = (this.windowRefService.nativeWindow as Window).location.pathname;
-
-        this.storeHelperService.dispatch(new Actions.UpdateURLPath(pathName));
     }
 }
