+++ web/src/main/webapp/v2/src/app/shared/components/form-field-error-message/form-field-error-message.component.css
@@ -1,9 +1,10 @@
 :host {
     display: block;
-    margin-top: 7px;
     padding-left: 7px;
 }
 
 .l-error-message {
+    display: block;
+    margin-top: 7px;
     color: #a94442;
 }
